# Remove allocations from randomChoice2 policy