--- a/pkg/activator/net/lb_policy.go
+++ b/pkg/activator/net/lb_policy.go
@@ -45,10 +45,9 @@ func randomChoice2Policy(_ context.Context, targets []*podTracker) (func(), *pod
 	l := len(targets)
 	// One tracker = no choice.
 	if l == 1 {
-		targets[0].addWeight(1)
-		return func() {
-			targets[0].addWeight(-1)
-		}, targets[0]
+		pick := targets[0]
+		cb := pick.increaseWeight()
+		return cb, pick
 	}
 	r1, r2 := 0, 1
 	// Two trackers - we know both contestants,
@@ -70,10 +69,8 @@ func randomChoice2Policy(_ context.Context, targets []*podTracker) (func(), *pod
 	if pick.getWeight() > alt.getWeight() {
 		pick = alt
 	}
-	pick.addWeight(1)
-	return func() {
-		pick.addWeight(-1)
-	}, pick
+	cb := pick.increaseWeight()
+	return cb, pick
 }
 
 // firstAvailableLBPolicy is a load balancer policy, that picks the first target