<!--
Request Prow to automatically lint any go code in this PR:

/lint
-->

<!-- Please include the 'why' behind your changes if no issue exists -->
## Proposed Changes

This is run as part of **each** request that has CC=0. Might as well make that as quick and cheap as we can.

```
benchmark                                                               old ns/op     new ns/op     delta
BenchmarkPolicy/random-power-of-2-choice-1-trackers-sequential-16       50.4          10.1          -79.96%
BenchmarkPolicy/random-power-of-2-choice-1-trackers-parallel-16         35.9          24.2          -32.59%
BenchmarkPolicy/random-power-of-2-choice-2-trackers-sequential-16       34.8          9.84          -71.72%
BenchmarkPolicy/random-power-of-2-choice-2-trackers-parallel-16         38.8          40.4          +4.12%
BenchmarkPolicy/random-power-of-2-choice-3-trackers-sequential-16       63.1          34.9          -44.69%
BenchmarkPolicy/random-power-of-2-choice-3-trackers-parallel-16         487           410           -15.81%
BenchmarkPolicy/random-power-of-2-choice-10-trackers-sequential-16      70.2          40.4          -42.45%
BenchmarkPolicy/random-power-of-2-choice-10-trackers-parallel-16        537           447           -16.76%
BenchmarkPolicy/random-power-of-2-choice-100-trackers-sequential-16     69.1          39.2          -43.27%
BenchmarkPolicy/random-power-of-2-choice-100-trackers-parallel-16       550           492           -10.55%

benchmark                                                               old allocs     new allocs     delta
BenchmarkPolicy/random-power-of-2-choice-1-trackers-sequential-16       1              0              -100.00%
BenchmarkPolicy/random-power-of-2-choice-1-trackers-parallel-16         1              0              -100.00%
BenchmarkPolicy/random-power-of-2-choice-2-trackers-sequential-16       1              0              -100.00%
BenchmarkPolicy/random-power-of-2-choice-2-trackers-parallel-16         1              0              -100.00%
BenchmarkPolicy/random-power-of-2-choice-3-trackers-sequential-16       1              0              -100.00%
BenchmarkPolicy/random-power-of-2-choice-3-trackers-parallel-16         1              0              -100.00%
BenchmarkPolicy/random-power-of-2-choice-10-trackers-sequential-16      1              0              -100.00%
BenchmarkPolicy/random-power-of-2-choice-10-trackers-parallel-16        1              0              -100.00%
BenchmarkPolicy/random-power-of-2-choice-100-trackers-sequential-16     1              0              -100.00%
BenchmarkPolicy/random-power-of-2-choice-100-trackers-parallel-16       1              0              -100.00%

benchmark                                                               old bytes     new bytes     delta
BenchmarkPolicy/random-power-of-2-choice-1-trackers-sequential-16       32            0             -100.00%
BenchmarkPolicy/random-power-of-2-choice-1-trackers-parallel-16         32            0             -100.00%
BenchmarkPolicy/random-power-of-2-choice-2-trackers-sequential-16       16            0             -100.00%
BenchmarkPolicy/random-power-of-2-choice-2-trackers-parallel-16         16            0             -100.00%
BenchmarkPolicy/random-power-of-2-choice-3-trackers-sequential-16       16            0             -100.00%
BenchmarkPolicy/random-power-of-2-choice-3-trackers-parallel-16         16            0             -100.00%
BenchmarkPolicy/random-power-of-2-choice-10-trackers-sequential-16      16            0             -100.00%
BenchmarkPolicy/random-power-of-2-choice-10-trackers-parallel-16        16            0             -100.00%
BenchmarkPolicy/random-power-of-2-choice-100-trackers-sequential-16     16            0             -100.00%
BenchmarkPolicy/random-power-of-2-choice-100-trackers-parallel-16       16            0             -100.00%
```

**Release Note**

<!-- Enter your extended release note in the below block. If the PR requires
additional action from users switching to the new release, include the string
"action required". If no release note is required, write "NONE". -->

```release-note
NONE
```

/assign @julz @vagababov